Use the MemorySegment scorer on JDK 21 with appropriate modifications
